* skill, code 정리.

+++ framework-cs/apps/UnityExample/Assets/Samples/Devian Samples/0.1.0/MobileSystem/Runtime/RewardManager/RewardManager.cs
@@ -43,7 +43,7 @@ namespace Devian
             if (string.IsNullOrEmpty(itemId) || count <= 0)
                 return;
 
-            list.Add(new RewardData(RewardType.Item, itemId, (long)count));
+            list.Add(new RewardData(RewardType.Item, itemId, count));
         }
     }
 }
